% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correlation_distance}
\alias{correlation_distance}
\title{Find the correlation distance between two vectors}
\usage{
correlation_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\description{
Find the correlation distance between two vectors
}
