% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Find the cosine distance between two vectors}
\usage{
cosine_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\description{
Find the cosine distance between two vectors
}
