% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Find the Euclidean (L2) distance between two vectors}
\usage{
euclidean_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\description{
Find the Euclidean (L2) distance between two vectors
}
