% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Find the Hamming distance between two vectors}
\usage{
hamming_distance(x, y)
}
\arguments{
\item{x}{An integer vector.}

\item{y}{An integer vector of the same length as \code{x}.}
}
\description{
Find the Hamming distance between two vectors
}
