% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l2sqr_distance}
\alias{l2sqr_distance}
\title{Find the squared Euclidean (squared L2) distance between two vectors}
\usage{
l2sqr_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\description{
Find the squared Euclidean (squared L2) distance between two vectors
}
