% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{Find the Manhattan (L1) distance between two vectors}
\usage{
manhattan_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\description{
Find the Manhattan (L1) distance between two vectors
}
