% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnndescent.R
\name{nn_brute_force}
\alias{nn_brute_force}
\title{Calculate exact nearest neighbors by brute force.}
\usage{
nn_brute_force(data, k, metric = "euclidean", n_threads = 0,
  grain_size = 1, verbose = FALSE)
}
\arguments{
\item{data}{Matrix of \code{n} items to generate random neighbors for.}

\item{k}{Number of nearest neighbors to return.}

\item{metric}{Type of distance calculation to use. One of \code{"euclidean"},
\code{"l2"} (squared Euclidean), \code{"cosine"}, \code{"manhattan"}
or \code{"hamming"}.}

\item{n_threads}{Number of threads to use.}

\item{grain_size}{Minimum batch size for multithreading. If the number of
items to process in a thread falls below this number, then no threads will
be used. Ignored if \code{n_threads < 1}.}

\item{verbose}{If \code{TRUE}, log information to the console.}
}
\value{
a list containing:
\itemize{
  \item \code{idx} an n by k matrix containing the nearest neighbor
  indices.
  \item \code{dist} an n by k matrix containing the nearest neighbor
   distances.
}
}
\description{
Calculate exact nearest neighbors by brute force.
}
