% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{nn_to_sparse}
\alias{nn_to_sparse}
\title{Convert Neighbor Graph to Sparse Distance Matrix}
\usage{
nn_to_sparse(nn)
}
\arguments{
\item{nn}{A nearest neighbor graph.}
}
\value{
the data in \code{nn} as a sparse distance matrix in \code{dgCMatrix} format.
}
\description{
Distances are stored column-wise, i.e. the neighbors of the first
observation in \code{nn} are in the first column of the matrix, neighbors of the
second observation are in the second column, and so on.
}
\details{
Zero distances are dropped. In the typical (non-bipartite) graph case, an
observation is usually a neighbor of itself with a distance of zero. These
distances are \emph{not} retained in the output, and hence if the number of
neighbors in \code{nn} is \code{k}, only \code{k - 1} neighbors are stored in the sparse
matrix.
}
\examples{
# find 4 nearest neighbors of first ten iris data
i10nn <- brute_force_knn(iris[1:10, ], k = 4)
# Nearest neighbors of each item is itself
all(i10nn$idx[, 1] == 1:10) # TRUE
# Convert to sparse
i10nnsp <- nn_to_sparse(i10nn)
# 3 neighbors are retained in the sparse format because we drop 0 distances
all(diff(i10nnsp@p) == 3) # TRUE
}
