% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnndescent.R
\name{nnd_knn_query}
\alias{nnd_knn_query}
\title{Find Nearest Neighbors and Distances}
\usage{
nnd_knn_query(
  reference,
  reference_nn,
  query,
  k = NULL,
  metric = "euclidean",
  init = NULL,
  n_iters = Inf,
  epsilon = 0.1,
  max_candidates = NULL,
  use_alt_metric = TRUE,
  n_threads = 0,
  grain_size = 1,
  verbose = FALSE
)
}
\arguments{
\item{reference}{Matrix of \code{m} reference items. The nearest neighbors to the
queries are calculated from this data.}

\item{reference_nn}{Nearest neighbors for the \code{reference} data, for
  example, the output of running `nnd_knn`. The format is a list containing:
\itemize{
  \item \code{idx} an n by k matrix containing the nearest neighbor indices
  of the data in \code{reference}.
  \item \code{dist} an n by k matrix containing the nearest neighbor
  distances.
}}

\item{query}{Matrix of \code{n} query items.}

\item{k}{Number of nearest neighbors to return. Optional if \code{init} is
specified.}

\item{metric}{Type of distance calculation to use. One of \code{"euclidean"},
\code{"l2sqr"} (squared Euclidean), \code{"cosine"}, \code{"manhattan"},
\code{"correlation"} (1 minus the Pearson correlation), or
\code{"hamming"}.}

\item{init}{Initial data to optimize. If not provided, \code{k} random
  neighbors are created. The input format should be the same as the return
  value: a list containing:
\itemize{
  \item \code{idx} an n by k matrix containing the nearest neighbor indices
  of the data in \code{reference}.
  \item \code{dist} an n by k matrix containing the nearest neighbor
  distances.
}
If \code{k} and \code{init} are provided then \code{k} must be equal to or
smaller than the number of neighbors provided in \code{init}. If smaller,
only the \code{k} closest value in \code{init} are retained. The input
distances may be ignored if \code{use_alt_metric = TRUE} and no
transformation from the distances to the internal distance representation
is available. In this case, the distances will be recalculated internally
and only the contents of \code{init$idx} will be used.}

\item{n_iters}{Number of iterations of nearest neighbor descent to carry out.
This is set to \code{Inf} by default. For controlling the time cost of
querying, it is recommended to modify \code{epsilon} initially. However,
setting this parameter can also provide a safe-guard against excessive
search time.}

\item{epsilon}{Controls trade-off between accuracy and search cost, by
specifying a distance tolerance on whether to explore the neighbors of
candidate points. The larger the value, the more neighbors will be
searched. A value of 0.1 allows query-candidate distances to be 10% larger
than the current most-distant neighbor of the query point, 0.2 means 20%,
and so on. Suggested values are between 0-0.5, although this value is
highly dependent on the distribution of distances in the dataset (higher
dimensional data should choose a smaller cutoff). Too large a value of
\code{epsilon} will result in the query search approaching brute force
comparison. Use this parameter in conjunction with \code{n_iters} and
\code{max_candidates} to prevent excessive run time. Default is 0.1.}

\item{max_candidates}{Maximum number of candidate neighbors to try for each
item in each iteration. Use relative to \code{k} to emulate the "rho"
sampling parameter in the nearest neighbor descent paper. By default, this
is set to \code{k} or \code{60}, whichever is smaller.}

\item{use_alt_metric}{If \code{TRUE}, use faster metrics that maintain the
ordering of distances internally (e.g. squared Euclidean distances if using
\code{metric = "euclidean"}), then apply a correction at the end. Probably
the only reason to set this to \code{FALSE} is if you suspect that some
sort of numeric issue is occurring with your data in the alternative code
path.}

\item{n_threads}{Number of threads to use.}

\item{grain_size}{Minimum batch size for multithreading. If the number of
items to process in a thread falls below this number, then no threads will
be used. Ignored if \code{n_threads < 1}.}

\item{verbose}{If \code{TRUE}, log information to the console.}
}
\value{
a list containing:
\itemize{
  \item \code{idx} an n by k matrix containing the nearest neighbor indices.
  \item \code{dist} an n by k matrix containing the nearest neighbor
   distances.
}
}
\description{
Find Nearest Neighbors and Distances
}
\examples{
# 100 reference iris items
iris_ref <- iris[iris$Species \%in\% c("setosa", "versicolor"), ]

# 50 query items
iris_query <- iris[iris$Species == "versicolor", ]

# First, find the approximate 4-nearest neighbor graph for the references:
iris_ref_knn <- nnd_knn(iris_ref, k = 4)

# For each item in iris_query find the 4 nearest neighbors in iris_ref.
# You need to pass both the reference data and the knn graph indices (the
# 'idx' matrix in the return value of nnd_knn).
# If you pass a data frame, non-numeric columns are removed.
# set verbose = TRUE to get details on the progress being made
iris_query_nn <- nnd_knn_query(iris_ref, iris_ref_knn, iris_query,
  k = 4, metric = "euclidean",
  verbose = TRUE
)
}
\references{
Dong, W., Moses, C., & Li, K. (2011, March).
Efficient k-nearest neighbor graph construction for generic similarity measures.
In \emph{Proceedings of the 20th international conference on World Wide Web}
(pp. 577-586).
ACM.
\url{https://doi.org/10.1145/1963405.1963487}.
}
