% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnndescent.R
\name{rnnd_prepare}
\alias{rnnd_prepare}
\title{Prepare approximate nearest neighbors index for querying}
\usage{
rnnd_prepare(index, n_threads = 0, verbose = FALSE)
}
\arguments{
\item{index}{An approximate nearest neighbor index produced by \code{\link[=rnnd_build]{rnnd_build()}}.}

\item{n_threads}{Number of threads to use.}

\item{verbose}{If \code{TRUE}, log information to the console.}
}
\value{
the approximate nearest neighbor index, prepared for querying.
}
\description{
Takes a nearest neighbor index produced by \code{rnnd_build} and uses the graph
diversification method of Harwood and Drummond (2016) to prepare a nearest
neighbor index for use in querying with \code{\link[=rnnd_query]{rnnd_query()}}.
}
\details{
If you ran \code{\link[=rnnd_build]{rnnd_build()}} with \code{prepare = TRUE} you do not need to run this
function before calling \code{\link[=rnnd_query]{rnnd_query()}}. If you run \code{\link[=rnnd_query]{rnnd_query()}} on an
index which has not been prepared, it will be prepared before the query is
run.

Calling \code{rnnd_prepare} on an already prepared index does nothing.
}
\examples{
iris_even <- iris[seq_len(nrow(iris)) \%\% 2 == 0, ]
iris_odd <- iris[seq_len(nrow(iris)) \%\% 2 == 1, ]

# Build index for knn only
iris_even_index <- rnnd_build(iris_even, k = 4)

# Prepare it for querying
iris_even_index <- rnnd_prepare(iris_even_index)

# Query new data
iris_odd_nbrs <- rnnd_query(index = iris_even_index, query = iris_odd, k = 4)

}
\references{
Harwood, B., & Drummond, T. (2016).
Fanng: Fast approximate nearest neighbour graphs.
In \emph{Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition}
(pp. 5713-5722).
}
\seealso{
\code{\link[=rnnd_build]{rnnd_build()}}, \code{\link[=rnnd_query]{rnnd_query()}}
}
