% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squared_euclidean_distance}
\alias{squared_euclidean_distance}
\title{Find the squared Euclidean (squared L2) distance between two vectors}
\usage{
squared_euclidean_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\description{
Find the squared Euclidean (squared L2) distance between two vectors
}
