% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnndescent.R
\name{random_nbrs}
\alias{random_nbrs}
\title{Randomly select nearest neighbors.}
\usage{
random_nbrs(data, k, metric = "euclidean", use_cpp = FALSE,
  n_threads = 0)
}
\arguments{
\item{data}{Matrix of \code{n} items to generate random neighbors for.}

\item{k}{Number of nearest neighbors to return.}

\item{metric}{Type of distance calculation to use. One of \code{"euclidean"},
\code{"l2"} (squared Euclidean), \code{"cosine"}, \code{"manhattan"}
or \code{"hamming"}.}

\item{use_cpp}{If \code{TRUE}, use the faster C++ code path.}

\item{n_threads}{Number of threads to use. Ignored if \code{use_cpp = FALSE}.}
}
\value{
a list containing:
\itemize{
  \item \code{idx} an n by k matrix containing the nearest neighbor
  indices.
  \item \code{dist} an n by k matrix containing the nearest neighbor
   distances.
}
}
\description{
Randomly select nearest neighbors.
}
